refactor(frr): extract BFD accumulator to sibling file
Move the BFD profile/peer dedup machinery out of policy_render.go into a
cohesive sibling bfd.go (#6424). The bfdSection accumulator lets the
protocol renderer collect BFD profiles/peers across the default instance
and every VRF and emit a single global `bfd` block (#2550); it is a
self-contained unit unrelated to the surrounding policy rendering.

Moved whole, unchanged: the bfdProfile / bfdPeer / bfdSection types and
their methods (newBFDSection, addProfile, addPeer, empty, render) plus
bfdProfileName. No signatures, logic, or render output change; the
emitted frr.conf is byte-identical.

Validation: go build ./pkg/frr, gofmt, go vet clean. policy_render.go
drops from 1,522 to 1,398 LOC.
